--- .babelrc
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    ["@babel/preset-env",
-    {
-      "useBuiltIns": "entry",
-      "corejs": "3.17"
-    }]
-  ],
-  "plugins": [
-    "@babel/plugin-proposal-class-properties"
-  ]
-}
