+++ Pipfile
@@ -1,0 +1,20 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+aiosmtpd = "*"
+tornado = "*"
+urlextract = "*"
+bleach = "*"
+beautifulsoup4 = "*"
+py-friendlywords = "*"
+
+[dev-packages]
+pytest = "*"
+pytest-asyncio = "*"
+requests = "*"
+
+[requires]
+python_version = "3.11"
