+++ mailboxzero/config.py
@@ -1,0 +1,8 @@
+DEFAULTS = {
+    "domains": "mb0.wtte.ch,qmq.ch",
+    "max_email_age": 600,
+    "http_host": "127.0.0.1",
+    "http_port": "8880",
+    "smtp_port": 25,
+    "smtp_hostname": "mail.mb0.wtte.ch"
+}
