--- package.json
@@ -1,42 +1,0 @@
-{
-  "name": "mailbox0",
-  "version": "0.0.1",
-  "description": "A simple smtp and webserver for throwaway email addresses",
-  "main": "index.js",
-  "repository": "git@github.com:betatim/mailbox0.git",
-  "author": "Tim Head <betatim@gmail.com>",
-  "license": "BSD-3",
-  "private": false,
-  "scripts": {
-    "webpack": "webpack",
-    "webpack:watch": "webpack --watch",
-    "webpack:analyze": "webpack --profile --json"
-  },
-  "browserslist": "last 2 version, not dead, not IE 11, >0.5% and not IE 11",
-  "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/plugin-proposal-class-properties": "^7.13.0",
-    "@babel/preset-env": "^7.14.4",
-    "@github/time-elements": "^3.1.2",
-    "@hotwired/stimulus": "^3.0.0",
-    "@hotwired/stimulus-webpack-helpers": "^1.0.1",
-    "@hotwired/turbo": "^7.0.0",
-    "@popperjs/core": "^2.10.1",
-    "autoprefixer": "^10.3.7",
-    "babel-loader": "^8.2.2",
-    "bootstrap": "^5.0.1",
-    "core-js": "^3.17.3",
-    "css-loader": "^5.2.6",
-    "local-time": "^2.1.0",
-    "mini-css-extract-plugin": "^1.6.0",
-    "postcss": "^8.3.6",
-    "postcss-loader": "^6.2.0",
-    "purgecss-webpack-plugin": "^4.0.3",
-    "sass": "^1.34.0",
-    "sass-loader": "^12.0.0",
-    "style-loader": "^2.0.0",
-    "webpack": "^5.38.1",
-    "webpack-bundle-analyzer": "^4.4.2",
-    "webpack-cli": "^4.7.0"
-  }
-}
