+++ static/controllers/clipboard_controller.js
