+++ static/controllers/refresh_controller.js
