+++ static/index.js
